% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adj_Per_Cap_Spend.R
\name{Adj_Per_Cap_Spend}
\alias{Adj_Per_Cap_Spend}
\title{Calculates the Adjusted per Capita Spending Value}
\usage{
Adj_Per_Cap_Spend(
  Total_Grocery_Sales = 8.11541e+11,
  Total_Pop = 334233854,
  Households = 131200000,
  Weekly_Household_Spending = 120,
  Cumulative_Price_Inc = NULL,
  State_Index = 100
)
}
\arguments{
\item{Total_Grocery_Sales}{The total volume of grocery sales in the US. The default
is set to 811,541,000,000 USD taken from IBIS World.}

\item{Total_Pop}{Total US population. The default is set to 334,233,854 for the
year of 2022.}

\item{Households}{Total number of US households. The default is at 131,200,000 based
based on statista.com.}

\item{Weekly_Household_Spending}{Average weekly spending per household. The default
is set to 120 USD based on Vertical IQ}

\item{Cumulative_Price_Inc}{Estimated cumulative price increase. Value should be entered
as an integer.}

\item{State_Index}{Pricing index for the state. Value should be entered as an
integer.}
}
\description{
Calculates the adjusted per capita grocery spending based on estimated cumulative price
increases nationally and for a given state.
}
\author{
Harun Celik
}
