% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/County_Pct.R
\name{County_Pct}
\alias{County_Pct}
\title{Calculates the Percentage of Market Area Inside of the Given County}
\usage{
County_Pct(county_area, market_total_area)
}
\arguments{
\item{county_area}{total square mileage of the county.}

\item{market_total_area}{total square mileage of the market area.}
}
\description{
Calculates the percentage of the market size in relation to the county.
Market area is divided by the county area if the market size area is smaller than county area.
}
\author{
Harun Celik
}
