% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Depreciation.R
\name{Depreciation}
\alias{Depreciation}
\title{Returns Annual Costs for Depreciating Assets}
\usage{
Depreciation(scenario, depreciation_table)
}
\arguments{
\item{scenario}{scenario choices to determine which depreciation costs to calculate.
Input is a character of  `one`, `two`, or `three`.}

\item{depreciation_table}{a data frame of the depreciation values containing the cost and life
use values used for calculating annual depreciation.}
}
\description{
This function returns a data frame of depreciation values for costs related to different scenarios of costs.
The returned values include annual depreciation values
for Building or Remodeling, Leasehold Improvements, Parking Lot Improvements,
Shelving and Check Out Counters, Computer Equipment and POS
Vehicles, Display Cases (frozen or refrigerated), Refrigeration, Freezers, Meat Cutting Equipment,
and the Amortization of Start-Up Costs.
}
\author{
Harun Celik
}
