% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expense.R
\name{Expense}
\alias{Expense}
\title{Returns Expense Calculations for Estimating Total Costs}
\usage{
Expense(total_estimated_revenue, expense_table)
}
\arguments{
\item{total_estimated_revenue}{the total estimated revenue calculated in the revenue estimation}

\item{expense_table}{a data frame containing percentages to be used to calculate the cost values for expenses}
}
\description{
Calculates the expense values for the Cost of Goods Sold, Gross Margin,
Officer Compensation, Salaries, Rent, Other Operating Expense, Depreciation,
Total Expense, Operating Income/(Loss), Interest Expense, Other Income,
Pre-Tax Net Profit/(Loss)
}
\author{
Harun Celik
}
