% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interest.R
\name{Interest}
\alias{Interest}
\title{Returns Interest Calculations for Loans}
\usage{
Interest(interest_table)
}
\arguments{
\item{interest_table}{a data frame of the loan and interest values for calculating interest amount on loans.}
}
\description{
This function returns a data frame of loans with calculated interest values.
}
\author{
Harun Celik
}
