% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Market_Populations.R
\name{Market_Populations}
\alias{Market_Populations}
\title{Returns Calculated Populations for a Determined Market Area}
\usage{
Market_Populations(
  county_pct,
  metro_pop,
  county_pop,
  county_cities,
  non_county_cities
)
}
\arguments{
\item{county_pct}{percentage of market area inside of a county.
Value calculated by @seealso [County_Pct()].}

\item{metro_pop}{population of town in which the grocery store is in.}

\item{county_pop}{total population of county in which the grocery store is in.}

\item{county_cities}{a list of populations for cities within the county of the grocery store.}

\item{non_county_cities}{a list of populations for cities within the market area but
outside the county of the grocery store.}
}
\description{
This function returns a list of populations used for determining
the primary, secondary, and rare shoppers in a calculated market area.
}
\author{
Harun Celik
}
