% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadrant_Area.R
\name{Quadrant_Area}
\alias{Quadrant_Area}
\title{Calculates a market size for nearest grocery stores}
\usage{
Quadrant_Area(north_distance, east_distance, south_distance, west_distance)
}
\arguments{
\item{north_distance}{distance to northern store in miles.}

\item{east_distance}{distance to eastern store in miles.}

\item{south_distance}{distance to southern store in miles.}

\item{west_distance}{distance to western store in miles.}
}
\description{
This function calculates a market size area given distances (in miles)
to nearest grocery stores in four cardinal directions.
}
\author{
Harun Celik
}
