% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rent.R
\name{Rent}
\alias{Rent}
\title{Returns Annual Costs for Rent}
\usage{
Rent(rent_table)
}
\arguments{
\item{rent_table}{a data frame of the monthly rent values for calculating annual rent costs.}
}
\description{
This function returns a data frame of annual rent values. All rents are considered on
12 month periods.
}
\author{
Harun Celik
}
