% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shopper_Populations.R
\name{Shopper_Populations}
\alias{Shopper_Populations}
\title{Calculates Primary, Secondary, and Rare Shopper Counts}
\usage{
Shopper_Populations(
  populations_list,
  metro_primary_pct,
  metro_secondary_pct,
  metro_rare_pct,
  town_primary_pct,
  town_secondary_pct,
  town_rare_pct,
  rural_primary_pct,
  rural_secondary_pct,
  rural_rare_pct
)
}
\arguments{
\item{populations_list}{list of populations calculated by @seealso [Market_Populations()].}

\item{metro_primary_pct}{percentage value for calculating shoppers of metro populations.}

\item{metro_secondary_pct}{percentage value for calculating shoppers of metro populations.}

\item{metro_rare_pct}{percentage value for calculating shoppers of metro populations.}

\item{town_primary_pct}{percentage value for calculating shoppers of town populations.}

\item{town_secondary_pct}{percentage value for calculating shoppers of town populations.}

\item{town_rare_pct}{percentage value for calculating shoppers of town populations.}

\item{rural_primary_pct}{percentage value for calculating shoppers of rural populations.}

\item{rural_secondary_pct}{percentage value for calculating shoppers of rural populations.}

\item{rural_rare_pct}{percentage value for calculating shoppers of rural populations.}
}
\description{
calculates the population of primary, secondary and rare shoppers for the metro
town, and rural populations of the calculated market area.
}
\author{
Harun Celik
}
