% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Total_Revenue.R
\name{Total_Revenue}
\alias{Total_Revenue}
\title{Calculates Spending Based on Number of Shoppers and Percentage Spent on Groceries}
\usage{
Total_Revenue(
  Adjusted_Per_Capita,
  Primary_Pct,
  Secondary_Pct,
  Rare_Pct,
  Primary_Shoppers,
  Secondary_Shoppers,
  Rare_Shoppers
)
}
\arguments{
\item{Adjusted_Per_Capita}{dollar value of annual spending on groceries per capita
adjusted by CPI and state inflation.}

\item{Primary_Pct}{percentage of income spent by primary shoppers on groceries. Default set to 60 percent.}

\item{Secondary_Pct}{percentage of income spent by secondary shoppers on groceries. Default set to 25 percent.}

\item{Rare_Pct}{percentage of income spent by rare shoppers on groceries. Default set to 5 percent.}

\item{Primary_Shoppers}{total number of primary shoppers calculated in @seealso [Shopper_Populations()].}

\item{Secondary_Shoppers}{total number of secondary shoppers calculated in @seealso [Shopper_Populations()].}

\item{Rare_Shoppers}{total number of rare shoppers calculated in @seealso [Shopper_Populations()].}
}
\description{
calculates the amount of grocery spending for primary, secondary, and rare shoppers
through a provided percentage and aggregates the spending for a total estimated revenue value.
}
\author{
Harun Celik
}
